Split MCP lifecycle code along owner boundaries
Problem:
The completed MCP session migration exceeded repository capacity limits in several long-lived owner files and tests.

Root cause:
SDK adapters, session diagnostics, argument normalization, and deterministic fixtures were initially colocated with broader product surfaces.

Fix:
Extract the new responsibilities into focused owner files, group capability and auxiliary refresh state by lifetime, and keep all runtime behavior and public contracts unchanged.

Verification:
go run ./tools/repolint
go test ./...
go test -race ./internal/agent/... ./internal/plugin/... ./internal/boot/...
cd desktop && go test -race ./...
pnpm --dir desktop/frontend test:all
pnpm --dir desktop/frontend build
