+++ desktop/mcp_version.go
@@ -1,0 +1,7 @@
+package main
+
+import "reasonix/internal/plugin"
+
+func init() {
+	plugin.SetMCPClientVersion(version)
+}
