+++ internal/plugin/plugin_test_helpers_test.go
@@ -1,0 +1,12 @@
+package plugin
+
+import "reasonix/internal/tool"
+
+func findToolByName(tools []tool.Tool, name string) tool.Tool {
+	for _, candidate := range tools {
+		if candidate.Name() == name {
+			return candidate
+		}
+	}
+	return nil
+}
