+++ internal/plugin/tool_list_changed_test.go
@@ -130,7 +130,7 @@ func newControlledRefreshClient(t *testing.T, tr *controlledToolsTransport) (*Cl
 	t.Helper()
 	ctx, cancel := context.WithCancel(context.Background())
 	client := &Client{
-		name: "controlled", t: tr, spec: Spec{Name: "controlled"}, supportsToolListChanged: true,
+		name: "controlled", t: tr, spec: Spec{Name: "controlled"}, capabilities: clientCapabilities{toolsListChanged: true},
 		refresh: toolListRefreshState{ctx: ctx, cancel: cancel, wait: func(context.Context, time.Duration) error { return nil }},
 	}
 	tools, err := client.listTools(ctx)
@@ -616,10 +616,10 @@ func TestClientCloseCancelsBlockedToolListRefresh(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	tr := newNotificationToolsTransport()
 	client := &Client{
-		name:                    "blocked",
-		t:                       tr,
-		spec:                    Spec{Name: "blocked"},
-		supportsToolListChanged: true,
+		name:         "blocked",
+		t:            tr,
+		spec:         Spec{Name: "blocked"},
+		capabilities: clientCapabilities{toolsListChanged: true},
 		refresh: toolListRefreshState{
 			ctx:    ctx,
 			cancel: cancel,
